requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.8
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwcrypto==0.8
MarkupSafe==1.1.1
Naked==0.1.31
pycparser==2.20
pycryptodome==3.9.7
pyodbc==4.0.30
pyrsistent==0.17.3
python-jwt==3.3.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
